+++ client/src/components/Achievements/index.js
@@ -1,0 +1,3 @@
+import Achievements from "./Achievements";
+
+export default Achievements;
