--- client/src/components/Dashboard/index.js
@@ -1,3 +1,0 @@
-import Dashboard from './Dashboard';
-
-export default Dashboard;
