+++ client/src/components/Progress/index.js
@@ -1,0 +1,3 @@
+import Progress from './Progress';
+
+export default Progress;
