+++ client/src/components/Uploader/index.js
@@ -1,0 +1,3 @@
+import Uploader from './Uploader';
+
+export default Uploader;
