+++ client/src/components/UserInfo/index.js
@@ -1,0 +1,3 @@
+import UserInfo from './UserInfo';
+
+export default UserInfo;
